Make sure clock pulses have a defined length
Currently, clock pulses can be arbitrarily short, as the whole clock
delay is done after setting clk to low, setting the data bit, and
setting clk to high again. On fast processors, the clk pulse can become
too short to be recognized reliably, which may be the cause for
issue #5.

To fix that, split the delay in two equal halves, and do one of them
while the clock is low.

I duplicated all other calls to delay() to keep the delay lengths
stable. This might not be necessary, but shouldn't hurt either.